+++ wp-content/uploads/civicrm/ext/firewall/settings/firewall.setting.php
@@ -12,22 +12,6 @@
 use CRM_Firewall_ExtensionUtil as E;
 
 return [
-  'firewall_csrf_timeout' => [
-    'name' => 'firewall_csrf_timeout',
-    'type' => 'Integer',
-    'html_type' => 'text',
-    'default' => 43200,
-    'is_domain' => 1,
-    'is_contact' => 0,
-    'title' => E::ts('Firewall CSRF timeout (seconds)'),
-    'description' => E::ts('Time after which generated CSRF token expires (default 12 hours: 43200 seconds)'),
-    'html_attributes' => [],
-    'settings_pages' => [
-      'firewall' => [
-        'weight' => 10,
-      ]
-    ],
-  ],
   'firewall_reverse_proxy' => [
     'name' => 'firewall_reverse_proxy',
     'type' => 'Boolean',